实现man模块内部权限认证。实现member模块内部权限认证。调整各模块权限列表。增加对flashMessage的显示支持。

+++ modules/member/controllers/ManagerController.php
@@ -15,6 +15,7 @@ use yii\filters\VerbFilter;
 class ManagerController extends Controller
 {
     public $layout = 'manager';
+
     public function behaviors()
     {
         return [
@@ -30,9 +31,14 @@ class ManagerController extends Controller
     /**
      * Lists all User models.
      * @return mixed
+     * @permission member.users.list
      */
     public function actionIndex()
     {
+        if(!$this->checkAccess('member.users.list')){
+            return $this->redirect(['/man/default/info']);
+        }
+
         $dataProvider = new ActiveDataProvider([
             'query' => User::find(),
         ]);
@@ -46,9 +52,14 @@ class ManagerController extends Controller
      * Displays a single User model.
      * @param integer $id
      * @return mixed
+     * @permission member.users.view
      */
     public function actionView($id)
     {
+        if(!$this->checkAccess('member.users.view')){
+            return $this->redirect(['/man/default/info']);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -58,9 +69,14 @@ class ManagerController extends Controller
      * Creates a new User model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
+     * @permission member.users.create
      */
     public function actionCreate()
     {
+        if(!$this->checkAccess('member.users.create')){
+            return $this->redirect(['/man/default/info']);
+        }
+
         $model = new User();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -77,9 +93,13 @@ class ManagerController extends Controller
      * If update is successful, the browser will be redirected to the 'view' page.
      * @param integer $id
      * @return mixed
+     * @permission member.users.update
      */
     public function actionUpdate($id)
     {
+        if (!$this->checkAccess('member.users.update')) {
+            return $this->redirect(['/man/default/info']);
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -96,9 +116,13 @@ class ManagerController extends Controller
      * If deletion is successful, the browser will be redirected to the 'index' page.
      * @param integer $id
      * @return mixed
+     * @permission member.users.delete
      */
     public function actionDelete($id)
     {
+        if (!$this->checkAccess('member.users.delete')) {
+            return $this->redirect(['/man/default/info']);
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -118,6 +142,22 @@ class ManagerController extends Controller
         } else {
             throw new NotFoundHttpException('The requested page does not exist.');
         }
+    }
+
+    /**
+     * 判断当前登录管理员是否满足指定权限
+     * @param $permission string 指定权限名
+     * @return boolean
+     */
+    protected function checkAccess($permission)
+    {
+        /* @var $manager \yii\web\User */
+        $manager = \Yii::$app->manager;
+        if (!$manager->can($permission)) {
+            \Yii::$app->session->setFlash(\Yii::$app->params['flashMessageParam'], "您没有进行该操作的权限({$permission})!");
+            return false;
+        }
+        return true;
     }
 
 }
